Restructuring Kubernetes related components
Restructering the remaining components from the 'kubernetes' folder. Now
they are also in the resource folder.
