+++ src/components/resources/cluster/events/EventDetails.tsx
@@ -12,8 +12,8 @@ import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
 import { timeDifference } from '../../../../utils/helpers';
-import Metadata from '../../../kubernetes/Metadata';
-import Row from '../../../kubernetes/Row';
+import Metadata from '../../misc/template/Metadata';
+import Row from '../../misc/template/Row';
 
 export const eventSource = (value: V1EventSource): string => {
   if (value.host && value.component) {
