+++ src/components/resources/configAndStorage/persistentVolumeClaims/PersistentVolumeClaimDetails.tsx
@@ -6,12 +6,12 @@ import { V1PersistentVolumeClaim } from '@kubernetes/client-node'
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
-import Conditions from '../../../kubernetes/Conditions';
-import Configuration from '../../../kubernetes/Configuration';
-import List from '../../../kubernetes/List';
-import Metadata from '../../../kubernetes/Metadata';
-import Row from '../../../kubernetes/Row';
-import Status from '../../../kubernetes/Status';
+import List from '../../misc/List';
+import Conditions from '../../misc/template/Conditions';
+import Configuration from '../../misc/template/Configuration';
+import Metadata from '../../misc/template/Metadata';
+import Row from '../../misc/template/Row';
+import Status from '../../misc/template/Status';
 
 interface IPersistentVolumeClaimDetailsProps extends RouteComponentProps {
   item: V1PersistentVolumeClaim;
