+++ src/components/resources/configAndStorage/persistentVolumes/PersistentVolumeDetails.tsx
@@ -6,11 +6,11 @@ import { V1PersistentVolume } from '@kubernetes/client-node'
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
-import Configuration from '../../../kubernetes/Configuration';
-import List from '../../../kubernetes/List';
-import Metadata from '../../../kubernetes/Metadata';
-import Row from '../../../kubernetes/Row';
-import Status from '../../../kubernetes/Status';
+import List from '../../misc/List';
+import Configuration from '../../misc/template/Configuration';
+import Metadata from '../../misc/template/Metadata';
+import Row from '../../misc/template/Row';
+import Status from '../../misc/template/Status';
 
 interface IPersistentVolumeDetailsProps extends RouteComponentProps {
   item: V1PersistentVolume;
