+++ src/components/resources/discoveryAndLoadbalancing/services/ServiceDetails.tsx
@@ -8,11 +8,11 @@ import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
 import { matchLabels } from '../../../../utils/helpers';
-import Configuration from '../../../kubernetes/Configuration';
-import List from '../../../kubernetes/List';
-import Metadata from '../../../kubernetes/Metadata';
-import Row from '../../../kubernetes/Row';
-import Status from '../../../kubernetes/Status';
+import List from '../../misc/List';
+import Configuration from '../../misc/template/Configuration';
+import Metadata from '../../misc/template/Metadata';
+import Row from '../../misc/template/Row';
+import Status from '../../misc/template/Status';
 
 interface IServiceDetailsProps extends RouteComponentProps {
   item: V1Service;
