+++ src/components/resources/misc/podTemplate/Logs.tsx
@@ -18,9 +18,9 @@ import {
 import { close, ellipsisHorizontal, ellipsisVertical, list } from 'ionicons/icons';
 import React, { useContext, useEffect, useState } from 'react';
 
-import { IContext, TActivator } from '../../declarations';
-import { AppContext } from '../../utils/context';
-import Editor from '../misc/Editor';
+import { IContext, TActivator } from '../../../../declarations';
+import { AppContext } from '../../../../utils/context';
+import Editor from '../../../misc/Editor';
 
 const TAIL_LINES = 1000;
