+++ src/components/resources/misc/podTemplate/tolerations/Tolerations.tsx
@@ -10,7 +10,7 @@ import {
 import { V1Toleration } from '@kubernetes/client-node';
 import React from 'react';
 
-import IonCardEqualHeight from '../misc/IonCardEqualHeight';
+import IonCardEqualHeight from '../../../../misc/IonCardEqualHeight';
 
 interface ITolerationProps {
   tolerations: V1Toleration[];
