+++ src/components/resources/misc/podTemplate/volumes/Volume.tsx
@@ -10,26 +10,37 @@ import {
   IonTitle,
   IonToolbar,
 } from '@ionic/react';
-import { V1NodeAffinity, V1PodAffinity, V1PodAntiAffinity } from '@kubernetes/client-node'
+import { V1Volume } from '@kubernetes/client-node';
 import { close } from 'ionicons/icons';
 import yaml from 'js-yaml';
-import React, { useState } from 'react';
+import React, {useState} from 'react';
 
-import Editor from '../misc/Editor';
+import Editor from '../../../../misc/Editor';
 
-interface IAffinityProps {
-  title: string
-  affinity: V1NodeAffinity | V1PodAffinity | V1PodAntiAffinity;
+interface IVolumeProps {
+  volume: V1Volume;
 }
 
-const Affinity: React.FunctionComponent<IAffinityProps> = ({ title, affinity }) => {
+const Volume: React.FunctionComponent<IVolumeProps> = ({ volume }) => {
   const [showModal, setShowModal] = useState(false);
+
+  let kind = '';
+  let description = '';
+
+  for (let key of Object.keys(volume)) {
+    if (key && key !== 'name') {
+      kind = key[0].toUpperCase() + key.slice(1);
+      description = volume[key];
+      break;
+    }
+  }
 
   return (
     <React.Fragment>
-      <IonItem button={true} onClick={() => setShowModal(true)}>
+      <IonItem button={true} className="cursor-pointer" onClick={() => setShowModal(true)}>
         <IonLabel>
-          <h2>{title}</h2>
+          <h2>{volume.name}</h2>
+          <p>{kind}</p>
         </IonLabel>
       </IonItem>
 
@@ -41,15 +52,15 @@ const Affinity: React.FunctionComponent<IAffinityProps> = ({ title, affinity }) 
                 <IonIcon slot="icon-only" icon={close} />
               </IonButton>
             </IonButtons>
-            <IonTitle>{title}</IonTitle>
+            <IonTitle>{volume.name}</IonTitle>
           </IonToolbar>
         </IonHeader>
         <IonContent>
-          <Editor readOnly={true} value={yaml.safeDump(affinity)} fullHeight={true} />
+          <Editor readOnly={true} value={yaml.safeDump(description)} fullHeight={true} />
         </IonContent>
       </IonModal>
     </React.Fragment>
   )
 };
 
-export default Affinity;
+export default Volume;
