+++ src/components/resources/misc/podTemplate/volumes/Volumes.tsx
@@ -8,7 +8,7 @@ import {
 import { V1Volume } from '@kubernetes/client-node';
 import React from 'react';
 
-import IonCardEqualHeight from '../misc/IonCardEqualHeight';
+import IonCardEqualHeight from '../../../../misc/IonCardEqualHeight';
 import Volume from './Volume';
 
 interface IVolumesProps {
