+++ src/components/resources/misc/template/Configuration.tsx
@@ -7,14 +7,14 @@ import {
 } from '@ionic/react';
 import React from 'react';
 
-import IonCardEqualHeight from '../misc/IonCardEqualHeight';
+import IonCardEqualHeight from '../../../misc/IonCardEqualHeight';
 
-const Status: React.FunctionComponent = ({ children }) => {
+const Configuration: React.FunctionComponent = ({ children }) => {
   return (
     <IonCol sizeXs="12" sizeSm="12" sizeMd="12" sizeLg="6" sizeXl="6">
       <IonCardEqualHeight>
         <IonCardHeader>
-          <IonCardTitle>Status</IonCardTitle>
+          <IonCardTitle>Configuration</IonCardTitle>
         </IonCardHeader>
         <IonCardContent>
           <IonGrid>
@@ -26,4 +26,4 @@ const Status: React.FunctionComponent = ({ children }) => {
   )
 };
 
-export default Status;
+export default Configuration;
