+++ src/components/resources/misc/template/Data.tsx
@@ -10,37 +10,24 @@ import {
   IonTitle,
   IonToolbar,
 } from '@ionic/react';
-import { V1Volume } from '@kubernetes/client-node';
 import { close } from 'ionicons/icons';
-import yaml from 'js-yaml';
-import React, {useState} from 'react';
+import React, { useState } from 'react';
 
-import Editor from '../misc/Editor';
+import Editor from '../../../misc/Editor';
 
-interface IVolumeProps {
-  volume: V1Volume;
+interface IDataProps {
+  name: string;
+  data: string;
 }
 
-const Volume: React.FunctionComponent<IVolumeProps> = ({ volume }) => {
+const Data: React.FunctionComponent<IDataProps> = ({ name, data }) => {
   const [showModal, setShowModal] = useState(false);
-
-  let kind = '';
-  let description = '';
-
-  for (let key of Object.keys(volume)) {
-    if (key && key !== 'name') {
-      kind = key[0].toUpperCase() + key.slice(1);
-      description = volume[key];
-      break;
-    }
-  }
 
   return (
     <React.Fragment>
-      <IonItem button={true} className="cursor-pointer" onClick={() => setShowModal(true)}>
+      <IonItem button={true} onClick={() => setShowModal(true)}>
         <IonLabel>
-          <h2>{volume.name}</h2>
-          <p>{kind}</p>
+          <h2>{name}</h2>
         </IonLabel>
       </IonItem>
 
@@ -52,15 +39,15 @@ const Volume: React.FunctionComponent<IVolumeProps> = ({ volume }) => {
                 <IonIcon slot="icon-only" icon={close} />
               </IonButton>
             </IonButtons>
-            <IonTitle>{volume.name}</IonTitle>
+            <IonTitle>{name}</IonTitle>
           </IonToolbar>
         </IonHeader>
         <IonContent>
-          <Editor readOnly={true} value={yaml.safeDump(description)} fullHeight={true} />
+          <Editor readOnly={true} value={data} fullHeight={true} />
         </IonContent>
       </IonModal>
     </React.Fragment>
   )
 };
 
-export default Volume;
+export default Data;
