+++ src/components/resources/misc/template/Status.tsx
@@ -7,14 +7,14 @@ import {
 } from '@ionic/react';
 import React from 'react';
 
-import IonCardEqualHeight from '../misc/IonCardEqualHeight';
+import IonCardEqualHeight from '../../../misc/IonCardEqualHeight';
 
-const Configuration: React.FunctionComponent = ({ children }) => {
+const Status: React.FunctionComponent = ({ children }) => {
   return (
     <IonCol sizeXs="12" sizeSm="12" sizeMd="12" sizeLg="6" sizeXl="6">
       <IonCardEqualHeight>
         <IonCardHeader>
-          <IonCardTitle>Configuration</IonCardTitle>
+          <IonCardTitle>Status</IonCardTitle>
         </IonCardHeader>
         <IonCardContent>
           <IonGrid>
@@ -26,4 +26,4 @@ const Configuration: React.FunctionComponent = ({ children }) => {
   )
 };
 
-export default Configuration;
+export default Status;
