+++ src/components/resources/workloads/daemonSets/DaemonSetItem.tsx
@@ -6,7 +6,7 @@ import { V1DaemonSet } from '@kubernetes/client-node'
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
-import ItemStatus from '../../../kubernetes/ItemStatus';
+import ItemStatus from '../../misc/template/ItemStatus';
 
 interface IDaemonSetItemProps extends RouteComponentProps {
   item: V1DaemonSet;
