+++ src/components/resources/workloads/replicationControllers/ReplicationControllerItem.tsx
@@ -6,7 +6,7 @@ import { V1ReplicationController } from '@kubernetes/client-node'
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
-import ItemStatus from '../../../kubernetes/ItemStatus';
+import ItemStatus from '../../misc/template/ItemStatus';
 
 interface IReplicationControllerItemProps extends RouteComponentProps {
   item: V1ReplicationController;
