+++ src/components/resources/workloads/statefulSets/StatefulSetItem.tsx
@@ -6,7 +6,7 @@ import { V1StatefulSet } from '@kubernetes/client-node'
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
-import ItemStatus from '../../../kubernetes/ItemStatus';
+import ItemStatus from '../../misc/template/ItemStatus';
 
 interface IStatefulSetItemProps extends RouteComponentProps {
   item: V1StatefulSet;
